feat: Remove 'Run the Application' section from README
The 'Run the Application' section including the 'npm start' command has been removed from the README file. This change was made because the application startup process is likely to be different for every environment, and thus should not be pre-determined in the README. Users can determine the best way to start the application suitable for their specific use case and environment setup.

No related issues.