refactor: remove unused fetch function
The removal of the commented code is completed. The `getUserCampaigns` function appeared to be unused, hence was extraneous. The superfluous code was taking up space and potentially causing confusion, so its removal leads to clearer, more maintainable code. There seems to be no significant consequences from this removal as the function was commented out, which means it wasn't executing and thus was redundant.